# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meta Business Agent Platform",
    "version": "2.0.0",
    "description": "OpenAPI specification for the Meta Business Agent Platform APIs used by the WABIZAgent application. Covers onboarding, agent configuration (settings, business info, skills, knowledge files, websites, FAQs, allowlist), connectors and their tools, evaluation, agent test, agent events, and thread control. All endpoints under `/{entity_id}/...` accept `X-API-Version: 2.0.0`. The thread control endpoint under `/business/whatsapp/phone_numbers/...` uses `X-API-Version: 1.0.0`."
  },
  "servers": [
    {
      "url": "https://api.facebook.com"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/{entity_id}/agent_connectors": {
      "parameters": [
        {
          "$ref": "#/components/parameters/EntityId"
        },
        {
          "$ref": "#/components/parameters/ApiVersion2"
        }
      ],
      "get": {
        "tags": [
          "Connectors"
        ],
        "operationId": "listConnectors",
        "summary": "List all connectors for the entity.",
        "responses": {
          "200": {
            "description": "List of connectors.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Connector"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Connectors"
        ],
        "operationId": "createConnector",
        "summary": "Create a new connector.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectorRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Newly created connector.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connector"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "EntityId": {
        "in": "path",
        "name": "entity_id",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "The entity ID for the Meta Business Agent. Typically the WhatsApp Business Phone Number ID."
      },
      "ApiVersion2": {
        "in": "header",
        "name": "X-API-Version",
        "required": false,
        "schema": {
          "type": "string",
          "enum": [
            "2.0.0"
          ],
          "default": "2.0.0"
        },
        "description": "Agent Platform API version. Must be `2.0.0`."
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Standard error payload.",
        "required": [
          "title",
          "detail"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          }
        }
      },
...
      "ApiKeyParam": {
        "type": "object",
        "required": [
          "field_name",
          "value"
        ],
        "properties": {
          "field_name": {
            "type": "string",
            "description": "Header, query parameter, or body field name."
          },
          "value": {
            "type": "string",
            "description": "Secret value for this API key field."
          },
          "prefix": {
            "type": "string",
            "description": "Optional value prefix, such as `Bearer ` (with trailing space)."
          }
        }
      },
      "ApiKeyAuthConfig": {
        "type": "object",
        "description": "API key configuration. Provide when `auth_type` is `API_KEY`.",
        "properties": {
          "headers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKeyParam"
            }
          },
          "query_params": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKeyParam"
            }
          },
          "body_params": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKeyParam"
            }
          }
        }
      },
      "OAuth2ClientCredentialsAuthConfig": {
        "type": "object",
        "description": "OAuth 2.0 client credentials configuration.",
        "required": [
          "token_url",
          "scopes_to_request",
          "client_id",
          "client_secret"
        ],
        "properties": {
          "token_url": {
            "type": "string",
            "format": "uri"
          },
          "scopes_to_request": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "token_request_content_type": {
            "type": "string",
            "enum": [
              "application/x-www-form-urlencoded",
              "application/json"
            ],
            "default": "application/x-www-form-urlencoded"
          },
          "client_id": {
            "type": "string"
          },
          "client_secret": {
            "type": "string"
          }
        }
      },
      "UserAuthInjectionConfig": {
        "type": "object",
        "description": "Configuration for injecting user auth tokens into tool requests.",
        "required": [
          "location",
          "field_name"
        ],
        "properties": {
          "location": {
            "type": "string",
            "enum": [
              "headers",
              "query",
              "body",
              "path"
            ]
          },
          "field_name": {
            "type": "string"
          },
          "prefix": {
            "type": "string"
          }
        }
      },
      "ConnectionStatus": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "CONNECTED",
              "PENDING_OAUTH",
              "DISCONNECTED",
              "ERROR"
            ],
            "description": "Current connection status. Some deployments also emit `ACTIVE` / `EXPIRED`; treat unknown values as `ERROR`."
          },
          "error_message": {
            "type": "string"
          }
        }
      },
      "MtlsConfig": {
        "type": "object",
        "description": "mTLS certificate metadata (private key is never exposed).",
        "required": [
          "has_certificate"
        ],
        "properties": {
          "has_certificate": {
            "type": "boolean"
          },
          "fingerprint": {
            "type": "string"
          },
          "expires_at": {
            "type": "integer"
          },
          "subject": {
            "type": "string"
          },
          "client_certificate": {
            "type": "string"
          },
          "ca_certificate": {
            "type": "string"
          }
        }
      },
      "ConnectorAuthConfig": {
        "type": "object",
        "description": "Discriminated auth configuration for a connector.",
        "properties": {
          "api_key": {
            "$ref": "#/components/schemas/ApiKeyAuthConfig"
          },
          "oauth2_client_credentials": {
            "$ref": "#/components/schemas/OAuth2ClientCredentialsAuthConfig"
          }
        }
      },
      "ConnectorRequest": {
        "type": "object",
        "description": "Request body for creating or updating a connector.",
        "required": [
          "name",
          "description",
          "base_url",
          "auth_type"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name of the connector."
          },
          "description": {
            "type": "string",
            "description": "What this connector integrates with and its purpose."
          },
          "base_url": {
            "type": "string",
            "format": "uri",
            "description": "Base URL of the external API."
          },
          "auth_type": {
            "type": "string",
            "enum": [
              "API_KEY",
              "OAUTH2_CLIENT_CREDENTIALS",
              "MTLS"
            ],
            "description": "Authentication type. `MTLS` uses the `upsertCertificate` endpoint to provide the client certificate."
          },
          "auth_config": {
            "$ref": "#/components/schemas/ConnectorAuthConfig"
          },
          "user_auth_injection_config": {
            "$ref": "#/components/schemas/UserAuthInjectionConfig"
          },
          "requires_certificate": {
            "type": "boolean",
            "description": "Whether this connector requires mTLS client certificate."
          }
        }
      },
      "Connector": {
        "type": "object",
        "description": "A configured connector.",
        "required": [
          "id",
          "name",
          "description",
          "base_url",
          "auth_type",
          "connection_status"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "base_url": {
            "type": "string",
            "format": "uri"
          },
          "auth_type": {
            "type": "string",
            "enum": [
              "API_KEY",
              "OAUTH2_CLIENT_CREDENTIALS",
              "MTLS"
            ]
          },
          "auth_config": {
            "$ref": "#/components/schemas/ConnectorAuthConfig"
          },
          "mtls_config": {
            "$ref": "#/components/schemas/MtlsConfig"
          },
          "connection_status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "user_auth_injection_config": {
            "$ref": "#/components/schemas/UserAuthInjectionConfig"
          }
        }
      },
...
    }
  }
}